Get in hot water, all of you.

+++ raven/src/main/java/com/getsentry/raven/connection/RandomEventSampler.java
@@ -44,7 +44,7 @@ public class RandomEventSampler implements EventSampler {
      */
     @Override
     public boolean shouldSendEvent(Event event) {
-        double randomFloat = random.nextDouble();
-        return sampleRate > Math.abs(randomFloat);
+        double randomDouble = random.nextDouble();
+        return sampleRate > Math.abs(randomDouble);
     }
 }
